冠军奖杯 使用后不删除 兑换 FISH-125

+++ src/main/java/cn/chahuyun/economy/factory/PropFishUsageContext.java
@@ -152,7 +152,9 @@ public class PropFishUsageContext {
                 return;
             }
             abstractPropUsage.excute();
-            PluginManager.getPropsManager().deleteProp(userInfo, propsCard);
+            if(abstractPropUsage.getDeleteProp()){
+                PluginManager.getPropsManager().deleteProp(userInfo, propsCard);
+            }
         }
     }
